+++ src/templates/Author.tsx
@@ -50,6 +50,7 @@ const Div = styled.div`
 
 const SocialMediaDiv = styled(Div)`
   display: flex;
+  flex-direction: row;
   align-items: center;
   justify-content: center;
 `;