+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Colours of the images in the cover components to the new brand colours.
 - Colours of the images in the rule components to the new brand colours.
 
+### Fixed
+
+- Google Analytics toggle in the cookie settings.
+
 ## [0.22.0] - 2021-06-04
 
 ### Added